requirements.txt
torch==2.4.1+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
pydantic
fastapi
gunicorn
itsdangerous
Jinja2
MarkupSafe
Werkzeug
uvicorn
gunicorn
scikit-learn

Flask~=3.1.2
flask_sqlalchemy
psycopg2-binary
dotenv~=0.9.9
python-dotenv~=1.1.1
flask-cors~=6.0.1

ulid-py~=1.1.0
